# docs/api-docs.yaml
openapi: 3.1.0
info:
  title: EcoMóstoles B2B API
  description: API REST para el intercambio de residuos
  version: "1.0"
servers:
- url: https://localhost:8443
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /api/v1/demands/{id}:
    get:
      tags:
      - Demands
      summary: Get demand by ID
      description: Retrieves the complete detail of a single material exchange demand.
        Returns 404 if the demand does not exist.
      operationId: getDemandById
      parameters:
      - name: id
        in: path
        description: Database primary key of the demand
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: Demand found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DemandDTO"
        "404":
          description: Demand not found
        "500":
          description: Unexpected server error
    put:
      tags:
      - Demands
      summary: Update an existing demand
      description: "Replaces the editable fields of a demand. Company ownership, visit\
        \ count and publication dates are preserved. Returns 404 if the demand does\
        \ not exist."
      operationId: updateDemand
      parameters:
      - name: id
        in: path
        description: Database primary key of the demand to update
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        description: "New demand values (company, visits and publicationDate are ignored)"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemandDTO"
        required: true
      responses:
        "200":
          description: Demand updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DemandDTO"
        "400":
          description: Validation failed -- check request body
        "404":
          description: Demand not found
        "500":
          description: Unexpected server error
    delete:
      tags:
      - Demands
      summary: Delete a demand
      description: Permanently removes a demand from the platform. Returns 204 NO
        CONTENT on success. Returns 400 if the demand cannot be deleted (e.g. has
        associated agreements). Returns 404 if the demand does not exist.
      operationId: deleteDemand
      parameters:
      - name: id
        in: path
        description: Database primary key of the demand to delete
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "204":
          description: Demand deleted successfully
        "400":
          description: Cannot delete due to existing constraints
        "404":
          description: Demand not found
        "500":
          description: Unexpected server error
...
  /api/v1/demands:
    get:
      tags:
      - Demands
      summary: "List demands (paginated, optional keyword filter)"
      description: Returns a paginated list of material exchange demands. Provide
        `keyword` to filter by title or description (ACTIVE demands only). Omit `keyword`
        to retrieve all demands across all statuses.
      operationId: getAllDemands
      parameters:
      - name: keyword
        in: query
        description: Optional free-text keyword to search by title or description
          (returns only ACTIVE demands when provided)
        required: false
        schema:
          type: string
        example: madera
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 12
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "publicationDate,DESC"
          items:
            type: string
      responses:
        "200":
          description: Page of demands returned successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DemandDTO"
        "500":
          description: Unexpected server error
    post:
      tags:
      - Demands
      summary: Create a new demand
      description: "Publishes a new material exchange demand. Returns 201 CREATED\
        \ with a Location header pointing to the new resource. Fields 'id', 'publicationDate',\
        \ 'expiryDate', 'createdAt' and 'visits' are set server-side."
      operationId: createDemand
      requestBody:
        description: Demand data to publish (server-side fields are ignored)
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemandDTO"
        required: true
      responses:
        "201":
          description: Demand created successfully -- see Location header
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DemandDTO"
        "400":
          description: Validation failed -- check request body
        "500":
          description: Unexpected server error
...
  /api/v1/matches/demands/{demandId}:
    get:
      tags:
      - Smart Matching
      summary: Find optimal matches for a demand
      description: "Executes the heuristics engine to calculate compatibility scores\
        \ between a specific demand and the active market supply, returning the top\
        \ 5 results sorted by score."
      operationId: getBestMatches
      parameters:
      - name: demandId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Match results successfully calculated
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchResultDTO"
...
components:
  schemas:
    CompanyDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        commercialName:
          type: string
          minLength: 1
        contactEmail:
          type: string
          minLength: 1
        taxId:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        industrialSector:
          type: string
          minLength: 1
        description:
          type: string
          maxLength: 255
          minLength: 0
        role:
          type: string
        co2Saved:
          type: number
          format: double
        ranking:
          type: integer
          format: int32
        sector:
          type: string
        verified:
          type: boolean
        initial:
          type: string
        formattedCo2Saved:
          type: string
      required:
      - address
      - commercialName
      - contactEmail
      - description
      - industrialSector
      - phone
      - taxId
    OfferDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        wasteCategory:
          type: string
        quantity:
          type: number
          format: double
        unit:
          type: string
        price:
          type: number
          format: double
        availability:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - RESERVED
          - PAUSED
          - IN_NEGOTIATION
          - FINISHED
          - REPORTED
        publicationDate:
          type: string
          format: date-time
        visits:
          type: integer
          format: int32
        company:
          $ref: "#/components/schemas/CompanyDTO"
        closed:
          type: boolean
        category:
          type: string
        formattedPrice:
          type: string
        formattedWasteType:
          type: string
        formattedQuantity:
          type: string
        formattedPublicationDate:
          type: string
    DemandDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        wasteCategory:
          type: string
        description:
          type: string
        quantity:
          type: number
          format: double
        unit:
          type: string
        urgency:
          type: string
        maxBudget:
          type: number
          format: double
        pickupZone:
          type: string
        validity:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - PAUSED
          - CLOSED
        createdAt:
          type: string
          format: date-time
        publicationDate:
          type: string
          format: date-time
        expiryDate:
          type: string
          format: date-time
        visits:
          type: integer
          format: int32
        company:
          $ref: "#/components/schemas/CompanyDTO"
        formattedBudget:
          type: string
        daysRemaining:
          type: integer
          format: int64
        expired:
          type: boolean
        formattedWasteType:
          type: string
        formattedQuantity:
          type: string
        formattedPickupZone:
          type: string
        formattedValidity:
          type: string
        formattedPublicationDate:
          type: string
...
    MatchResultDTO:
      type: object
      properties:
        offer:
          $ref: "#/components/schemas/OfferDTO"
        matchScore:
          type: number
          format: double
        matchReason:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Introduce el token JWT obtenido en /api/v1/auth/login
      in: header
      scheme: bearer
      bearerFormat: JWT